feat(support): unify Donate + Commercial License behind one toggle
The donate page and the commercial-license (enterprise) page were two separate
full-screen modes reached from different places. Merge them into a single
SupportPage with a charming segmented toggle:

- Segmented "💛 Support ⇄ 🏢 Commercial License" control with a sliding active
  pill that carries each panel's accent hue (pink for Support, teal for
  License) and an icon that pops on selection.
- Switching cross-fades the panel (key remount replays the hero/card entry
  animations) over the shared Launchpad aurora + a single Back button.
- Both legacy modes still work: 'donate' opens the Support tab, 'enterprise'
  opens the Commercial License tab — so the footer heart and the dub/export
  "commercial license" links land on the right tab unchanged.

Reuses the existing donate/enterprise chrome (DonatePage.css + EnterprisePage.css
kept and imported); SupportPage.css only owns the toggle + transitions. Both
views share one 640px container width so the frame doesn't jump on toggle.
DonatePage.jsx + EnterprisePage.jsx removed (content folded in).

3 new support.* i18n keys, backfilled across 21 locales.
Verified: tsc clean, build OK (SupportPage chunk replaces the two old ones),
vitest 167/167, CJK guard passes.
